getTopTracks() now returns top song names+artists!
It returns it in the form of an array with Track Objects - I made a Track class
Open the console to see it in action when you press the buttons.
There is now two buttons, the first buttons fetch the top songs, it runs getTopTracks()
The second button just prints out the Top Tracks array to ensure that the App end sucessfully grabbed it. ie. it populated the topTracks app.jsx global variable correctly.