fixed login button, see extended comment. Fixes #5
The issue was that "import login from './App.jsx';" is not a valid thing to do in Login.jsx
FUNCTIONS CANNOT BE IMPORTED FROM A .jsx file like this!!
I had to pass the function in through app.jsx like this: <Login login={login} />
On another note, does login() need to be defined here? Or could it be defined
in a regular .js file to eliminate the need for this? If it were in a
typical .js file I assume the original code would have worked