I got the track images to display
I am not sure which change actually fixed it honestly, I was trying a lot of little things that didn't work and then suddenly it started working while I was trying to set up debugging to help me find the issue. My best guess is it was something with the variable names.