getTopTracks/Artis now limits self if... Fixes #12
getTopTracks/Artists now limits how much data it tried to fetch if needed. If a user does not have
enough listening history for the amount of data they request, getTopTracks() and getTopArtists()
will only try to fetch as much data as is available from the API call.
Previously if a user only had say 10 top tracks, but requested 20, the program would break
because it would try to force a fetch of 20 tracks. Now it will only go up to the amount available,
10 in this example.