"trends" page now shows similarity percentage
I also overhauled the structure of TasteComparator, and Trends.jsx
TasteComparator now only sets the national trends, so getting the
user tracks is handled on  the Trends.jsx page by it calling getTopTracks().
The tracks array used on Trends.jsx is also declared as a local variable,
so it will not interfere with the global tracks variable.

I also did some re-structureing on Trends.jsx to ensure that top user tracks are
fetched before the rest of the page loads in

slight modification on app.jsx to make sure it's call to Trends.jsx concurrs with the changes in Trends.jsx